nn_deploy/AndesAIRE-NN-SDK/requirements.txt
attrs==22.2.0
cloudpickle==2.2.1
decorator==5.1.1
GitPython==3.1.31
graphviz==0.20.1
matplotlib==3.7.0
netron
numba==0.56.4
numpy==1.23.5
onnx==1.12.0
opencv-python-headless==4.7.0.72
pandas==1.5.3
psutil==5.9.4
pyserial==3.5
pydot==1.4.2
pytorchcv==0.0.67
pyvww==0.1.1
PyYAML==6.0
scikit-learn==1.2.1
scipy==1.10.1
synr==0.6.0
tabulate==0.9.0
tensorboard==2.9.0
tensorflow==2.9.1
torchnet==0.0.4
tornado==6.2
tqdm==4.64.1
xgboost==1.7.4
librosa==0.8.1
pesq==0.0.4
soundfile==0.12.1
